feat: Refactor common components
